Fix emoji panic 😱
At the moment if an emoji is contained in the error message it gets
displayed and correctly truncated, but afterwards the bar is one too
short since the emoji takes less space visually then it takes in the
message length.
This can be fixed by lengthening the loading bar if the message is
displayed, but we have to add a few more things todo so.
TODO